backend/reservation_app/serializers.py
from rest_framework import serializers
from .models import Reservation

class ReservationSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Reservation
        fields = '__all__'