backend/requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.9.0